+++ tests/conftest.py
@@ -6,7 +6,7 @@ import pytest
 import yaml
 
 from user_sync.config.user_sync import UMAPIConfigLoader
-from .util import merge_dict, make_dict, ClearableStringIO
+from .util import merge_dict, make_dict, ClearableStringIO, create_blank_user
 
 
 @pytest.fixture
@@ -85,10 +85,12 @@ def modify_root_config(modify_config):
 def default_args(cli_args, test_resources):
     return cli_args({'config_filename': test_resources['umapi_root_config']})
 
+
 # A shortcut for loading the config file
 @pytest.fixture
 def default_sign_args(cli_args, test_resources):
     return cli_args({'config_filename': test_resources['sign_root_config']})
+
 
 @pytest.fixture
 def resource_file():
@@ -104,18 +106,43 @@ def resource_file():
 
     return _resource_file
 
-@pytest.fixture
-def example_user():
-    return {
-        'type': 'federatedID',
-        'username': 'user@example.com',
-        'domain': 'example.com',
-        'email': 'user@example.com',
-        'firstname': 'Example',
-        'lastname': 'User',
-        'groups': set(),
-        'country': 'US',
-    }
+
+@pytest.fixture()
+def get_mock_user():
+    def _get_mock_user(
+            identifier="user1",
+            is_umapi_user=False,
+            firstname=None,
+            lastname=None,
+            groups=None,
+            country="US",
+            identity_type="federatedID",
+            domain="example.com",
+            username=None
+    ):
+        u = create_blank_user(identifier, firstname, lastname, groups,
+                              country, identity_type, domain, username)
+        if is_umapi_user:
+            u.pop('identity_type')
+            u.pop('member_groups')
+            u.pop('source_attributes')
+        else:
+            u.pop('adminRoles')
+            u.pop('status')
+            u.pop('type')
+        return u
+
+    return _get_mock_user
+
+
+@pytest.fixture()
+def mock_dir_user(get_mock_user):
+    return get_mock_user()
+
+
+@pytest.fixture()
+def mock_umapi_user(get_mock_user):
+    return get_mock_user(is_umapi_user=True)
 
 
 @pytest.fixture
