+++ tests/util.py
@@ -48,6 +48,49 @@ def merge_dict(d1, d2, immutable=False):
             d1[k] = d2[k]
     return d1
 
+# Serves as a base user for either umapi or directory user tests
+def create_blank_user(
+        identifier,
+        firstname=None,
+        lastname=None,
+        groups=None,
+        country="US",
+        identity_type="federatedID",
+        domain="example.com",
+        username=None
+):
+    if '@' not in identifier:
+        identifier = identifier + "@" + domain
+    user_id = identifier.split("@")[0]
+    firstname = firstname or user_id + " First"
+    lastname = lastname or user_id + " Last"
+    domain = domain or identifier.split("@")[-1]
+
+    return deepcopy({
+        'identity_type': identity_type,
+        'type': identity_type,
+        'username': username or identifier,
+        'domain': domain,
+        'firstname': firstname,
+        'lastname': lastname,
+        'email': identifier,
+        'groups': groups or [],
+        'member_groups': [],
+        'adminRoles': [],
+        'status': 'active',
+        'country': country,
+        'source_attributes': {
+            'email': identifier,
+            'identity_type': None,
+            'username': user_id,
+            'domain': domain,
+            'givenName': firstname,
+            'sn': lastname,
+            'c': country}
+    })
+
+
+
 class MockResponse:
 
     def __init__(self, status=200, body=None, headers=None, text=None):
@@ -71,3 +114,5 @@ class ClearableStringIO(StringIO, object):
     def getvalue(self, *args, **kwargs):
         self.flush()
         return super(ClearableStringIO, self).getvalue()
+
+
